pyproject.toml
[tool.poetry]
name = "germinate-ai"
version = "0.1.0"
description = "An AI engine fAn AI engine for distributed multi-agent systems based on LLMs"
authors = ["bk62 <celeritatem@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
documentation = "https://github.com/bk62/germinate-ai"
repository = "https://github.com/bk62/germinate-ai"
keywords = ["AI", "LLM", "AI Agent", "Multi-agent System", "Distributed Computing"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries"
]
packages = [
    { include="germinate_ai" }
]

[tool.poetry.urls]
Homepage = "https://github.com/bk62/germinate-ai"
"Bug Tracker" = "https://github.com/bk62/germinate-ai/issues"

[tool.poetry.scripts]
germinate = "germinate_ai.__main__:cli"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.0"
loguru = "^0.7.2"
typer = "^0.12.3"
langchain-core = "^0.1.42"
langchain-community = "^0.0.32"
nats-py = "^2.7.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
langgraph = "^0.0.37"
beautifulsoup4 = "^4.12.3"
weaviate-client = "^4.5.5"
langchain-google-genai = "^1.0.2"
pillow = "^10.3.0"
langchain-weaviate = "^0.0.1.post1"
pydantic-settings = "^2.2.1"
networkx = "^3.3"
psycopg = "^3.1.18"
attrs = "^23.2.0"
cloudpickle = "^3.0.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pyright = "^1.1.358"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.20"
mkdocstrings = {extras = ["python"], version = "^0.25.0"}
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
coverage = "^7.4.4"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
pythonVersion = "3.11"
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint.isort]
known-first-party = ["germinate_ai"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E501", "W291", "W292", "W293"]
